Updated '@brunexx/ngx-universal' to 'ngxc-universal'
I found that the npm package is with the name "ngxc-universal" but in the README file name and directories were used as '@brunexx/ngx-universal' while there is no such directory or even package. So updated the README file to give readers a better and easy-to-use experience.
Server.ts file should import 'import { NgxRequest, NgxResponse } from 'ngxc-universal';' and use these values instead of using strings which eventually gives an error.