Strict typing for sameSite
SameSite can have one of three values which are 'Strict' | 'Lax' | 'None' whereas the type was string through which the user can pass any string value which eventually can cause the error. 
Note: These strings are used the same as the Readme file but these are a bit different from a main ngx-cookie library.